Created failed system
This was intended to be an update to allow for validation of the process. But because JSONPLaceholder is a Mock API, no data is actually stored, so none of the POST or PUT commands actually work. Meaning when I tried to get them, it didn't work. Still however, a valuable lesson to be learned